Move resources out of libraries to prevent resource duplication
